Db -> works Location -> works Location change listener -> OK Proximity -> WIP MAP -> WIP permissions works in a weird manner.

+++ app/src/main/java/com/example/s17149/DataBase/Shop.kt
@@ -9,10 +9,10 @@ data class Shop(
     var id: Long = 0,
     var name: String,
     var description: String,
-    var latitude: Long = 0L,
-    var longtitude: Long = 0L,
+    var latitude: Double = 0.0,
+    var longtitude: Double = 0.0,
     var favorite: Boolean = false,
-    var radius: Long = 0L,
+    var radius: Double = 0.0,
 ) {
 
 }
